Change: execute database upgrades from install script instead of via dashboard
