+++ src/admin/src/View/Cpanel/HtmlView.php
@@ -74,8 +74,6 @@ class HtmlView extends BaseHtmlView
 
         $this->KunenaMenusExists = KunenaMenuHelper::KunenaMenusExists();
 
-        $this->upgradeDatabase();
-
         $logFinder = new KunenaLogFinder();
         $this->numberOfLogs = $logFinder->count();
 
@@ -123,75 +121,6 @@ class HtmlView extends BaseHtmlView
         ToolbarHelper::spacer();
         $helpUrl = 'https://docs.kunena.org/en/';
         ToolbarHelper::help('COM_KUNENA', false, $helpUrl);
-    }
-
-    /**
-     * Method to upgrade the database at the end of installation.
-     *
-     * @return  boolean
-     *
-     * @throws Exception
-     * @since   Kunena 6.0
-     */
-    protected function upgradeDatabase()
-    {
-        $app = Factory::getApplication();
-
-        $xml = simplexml_load_file(JPATH_ADMINISTRATOR . '/components/com_kunena/install/kunena.install.upgrade.xml');
-
-        if ($xml === false) {
-            $app->enqueueMessage(Text::_('COM_KUNENA_INSTALL_DB_UPGRADE_FAILED_XML'), 'error');
-
-            return false;
-        }
-
-        // The column "state" in kunena_version indicate from which version to update
-        $db = Factory::getContainer()->get('DatabaseDriver');
-        $db->setQuery("SELECT state FROM #__kunena_version ORDER BY `id` DESC", 0, 1);
-        $stateVersion = $db->loadResult();
-
-        if (!empty($stateVersion)) {
-            $curversion = $stateVersion;
-        } else {
-            return false;
-        }
-
-        $modelInstall = new KunenaModelInstall();
-
-        foreach ($xml->upgrade[0] as $version) {
-            // If we have already upgraded to this version, continue to the next one
-            $vernum = (string) $version['version'];
-
-            if (!empty($status[$vernum])) {
-                continue;
-            }
-
-            // Update state
-            $status[$vernum] = 1;
-
-            if ($version['version'] == '@' . 'kunenaversion' . '@') {
-                $git    = 1;
-                $vernum = KunenaForum::version();
-            }
-
-            if (isset($git) || version_compare(strtolower($version['version']), strtolower($curversion), '>')) {
-                foreach ($version as $action) {
-                    $modelInstall->processUpgradeXMLNode($action);
-
-                    $app->enqueueMessage(Text::sprintf('COM_KUNENA_INSTALL_VERSION_UPGRADED', $vernum), 'success');
-                }
-
-                $query = "UPDATE `#__kunena_version` SET state='';";
-                $db->setQuery($query);
-
-                $db->execute();
-
-                // Database install continues
-                // return false;
-            }
-        }
-
-        return true;
     }
 
     /**
