+++ src/script.php
@@ -321,10 +321,6 @@ return new class() implements ServiceProviderInterface {
                         }
                     }
 
-                    if ($installed != 'NONE') {
-                        $state = $installed;
-                    }
-
                     $query = $db->createQuery();
 
                     $values = [
